LayoutTests/platform/glib/fast/mediastream/RTCPeerConnection-remoteDescription-expected.txt

Tests RTCPeerConnection remoteDescription.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.setRemoteDescription(sessionDescription).then(requestSucceeded1, requestFailed1); did not throw exception.
PASS requestSucceeded was called.
PASS pc.setRemoteDescription(sessionDescription).then(requestSucceeded2, requestFailed2); did not throw exception.
PASS requestFailed was called.
PASS pc.remoteDescription.type is "offer"
FAIL pc.remoteDescription.sdp should be remote. Was t=0 0
.
PASS pc.remoteDescription.type is "offer"
FAIL pc.remoteDescription.sdp should be remote. Was t=0 0
.
PASS successfullyParsed is true

TEST COMPLETE
